# contracts/catalog.openapi.yaml
openapi: 3.1.0
info:
  title: deedlit.catalog
  version: 0.1.0
  description: >
    Source-of-truth catalog over Postgres + RustFS. Images, tags, params,
    references, ratings, favorites, notes, collections, and blob I/O for
    thumbnails / cached embeddings. Owning service: never calls search or graph.
servers:
  - url: http://deedlit-catalog:8000
paths:
...
  /images/batch-delete:
    post:
      summary: Bulk hard-delete many image records (+ blobs); not the source files
      description: >
        The batch counterpart to DELETE /images/{sha256}: two set-based SQL deletes
        (asset refs, then the image rows — FK children cascade) plus per-blob
        object-store cleanup for exactly the records that existed. Returns the
        deleted + missing sha256s.
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { sha256s: { type: array, items: { $ref: "#/components/schemas/Sha256" }, minItems: 1, maxItems: 1000 } }, required: [sha256s] } } } }
      responses: { "200": { description: OK, content: { application/json: { schema: { type: object, properties: { status: { type: string }, deleted: { type: array, items: { $ref: "#/components/schemas/Sha256" } }, missing: { type: array, items: { $ref: "#/components/schemas/Sha256" } } } } } } } }
  /images/{sha256}:
    parameters: [ { name: sha256, in: path, required: true, schema: { $ref: "#/components/schemas/Sha256" } } ]
    get: { summary: Read one image, responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Image" } } } }, "404": { description: Not found } } }
    patch:
      summary: Update mutable fields (rating/favorite/tags)
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ImagePatch" } } } }
      responses: { "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Image" } } } } }
    delete:
      summary: Delete an image record (+ sha256-keyed blobs); not the source file
      description: >
        Hard-deletes the Postgres row (FK children cascade) plus the per-image
        asset references, then drops the sha256-keyed RustFS blobs (thumbnail +
        cached embedding). The original file on disk is never touched. Note /
        collection membership refs are left intact so a re-ingest re-attaches.
      responses:
        "200": { description: Deleted, content: { application/json: { schema: { type: object, properties: { status: { type: string } } } } } }
        "404": { description: Not found }
...
components:
  schemas:
...
    Sha256: { type: string, pattern: "^[a-f0-9]{64}$" }
...
    AssetRef: { type: object, properties: { kind: { type: string, enum: [checkpoint, lora, embedding, vae, controlnet, upscaler] }, name: { type: string }, hash: { type: [string, "null"] } }, required: [kind, name] }
    Params:
      type: object
      properties:
        seed: { type: [integer, "null"] }
        steps: { type: [integer, "null"] }
        cfg: { type: [number, "null"] }
        sampler: { type: [string, "null"] }
        scheduler: { type: [string, "null"] }
        denoise: { type: [number, "null"] }
        clipskip: { type: [integer, "null"] }
        width: { type: [integer, "null"] }
        height: { type: [integer, "null"] }
    ImageUpsert:
      type: object
      properties:
        sha256: { $ref: "#/components/schemas/Sha256" }
        filepath:
          type: [string, "null"]
          description: >-
            Original on-disk path of the source file captured at ingest, so an
            image is identifiable when inspecting the catalog (the id is the
            opaque sha256). Stored INSERT-only — a later reindex (which has no
            path) never overwrites it. Null when unknown.
        phash: { type: [string, "null"] }
        width: { type: [integer, "null"] }
        height: { type: [integer, "null"] }
        sourceTool: { type: [string, "null"] }
        prompt: { type: [string, "null"] }
        negative: { type: [string, "null"] }
        tags: { type: array, items: { type: string } }
        params: { $ref: "#/components/schemas/Params" }
        references: { type: array, items: { $ref: "#/components/schemas/AssetRef" } }
        workflow_json: { type: [object, "null"] }
        api_prompt_json: { type: [object, "null"] }
        safety:
          type: [string, "null"]
          enum: [sfw, nsfw, explicit, null]
          description: AI content-safety class (deedlit.labelagent); null = unclassified.
        description:
          type: [string, "null"]
          description: >-
            AI-generated image description (deedlit.labelagent). An expensive
            vision-LLM result, persisted in image_descriptions so it is
            retrievable/viewable without re-running the model. Re-derivable: a
            reindex without one keeps the stored value. Null = none.
      required: [sha256]
    Image:
      allOf:
        - { $ref: "#/components/schemas/ImageUpsert" }
        - type: object
          properties:
            rating: { type: [integer, "null"] }
            favorite: { type: boolean }
            created_at: { type: string, format: date-time }
...
    ImagePatch:
      type: object
      properties:
        rating: { type: [integer, "null"] }
        favorite: { type: [boolean, "null"] }
        tags: { type: [array, "null"], items: { type: string } }
        safety: { type: [string, "null"], enum: [sfw, nsfw, explicit, null] }
...